Block editor: scroll selected block only if it has no focus
