+++ packages/block-editor/src/components/block-list/use-block-props/index.js
@@ -29,6 +29,7 @@ import { useBlockMovingModeClassNames } from './use-block-moving-mode-class-name
 import { useEventHandlers } from './use-event-handlers';
 import { useNavModeExit } from './use-nav-mode-exit';
 import { useBlockNodes } from './use-block-nodes';
+import { useScrollIntoView } from './use-scroll-into-view';
 import { store as blockEditorStore } from '../../../store';
 
 /**
@@ -102,6 +103,8 @@ export function useBlockProps( props = {}, { __unstableIsHtml } = {} ) {
 	const mergedRefs = useMergeRefs( [
 		props.ref,
 		useFocusFirstElement( clientId ),
+		// Must happen after focus because we check for focus in the block.
+		useScrollIntoView( clientId ),
 		useBlockNodes( clientId ),
 		useEventHandlers( clientId ),
 		useNavModeExit( clientId ),
